Use shared EEs instead of local ones
So we can maintain a single implementation of the library EEs,
Architecture, Logging, and Time, I deleted these EEs from the
testbenching project, causing this component to use the EEs resident
within the WVBalloonAutopilot project.  Eventually, it is likely we will
put these things in a separate project so they can be shared across
multiple applications, but for now, it is convenient to leave them in
the WVAutopilot project.

Since Verifier does not like to source-locate for bridge operations
across project boundaries (https://support.onefact.net/issues/10302), I
also replaced the invocation of ARCH::shutdown with the equivalent OAL
primitive.